agregando validacion para producto, su traduccion y importamos componente globales
