+++ app/Models/OrderDetails.php
@@ -15,4 +15,9 @@ class OrderDetails extends Model
     {
         return $this->belongsTo(Order::class);
     }
+
+    public function products()
+    {
+        return $this->belongsTo(Product::class);
+    }
 }
