+++ resources/js/app.js
@@ -3,8 +3,16 @@ require('./bootstrap');
 import { createApp } from "vue";
 import router from './router';
 import App from "./App.vue";
+import DeleteComponent from "./components/DeleteComponent";
+import BreadcrumbComponent from "./components/BreadcrumbComponent";
+import PaginationComponent from "./components/PaginationComponent";
 import axios from "axios";
 import VueSweetalert2 from 'vue-sweetalert2';
 import 'sweetalert2/dist/sweetalert2.min.css';
 
-createApp(App).use(router,axios,VueSweetalert2).mount('#app');
+const app = createApp(App);
+
+app.component('DeleteComponent', DeleteComponent);
+app.component('BreadcrumbComponent', BreadcrumbComponent);
+app.component('PaginationComponent', PaginationComponent);
+app.use(router,axios,VueSweetalert2).mount('#app');
